type comparisons: fixes
This is a huge patch that should be a global NOP.

For unix type platforms it enables -Wconversion to issue warnings (-> error)
for all automatic casts that seem less than ideal but are normally concealed
by the toolchain.

This is things like passing an int to a size_t argument.  Once enabled, I
went through all args on my default build (which build most things) and
tried to make the removed default cast explicit.

With that approach it neither change nor bloat the code, since it compiles
to whatever it was doing before, just with the casts made explicit... in a
few cases I changed some length args from int to size_t but largely left
the causes alone.

From now on, new code that is relying on less than ideal casting
will complain and nudge me to improve it by warnings.

+++ lib/core-net/client/connect2.c
@@ -250,7 +250,7 @@ solo:
 
 	if (wsi->a.vhost->http.http_proxy_port) {
 		ads = wsi->a.vhost->http.http_proxy_address;
-		port = wsi->a.vhost->http.http_proxy_port;
+		port = (int)wsi->a.vhost->http.http_proxy_port;
 #else
 		if (0) {
 #endif
@@ -262,7 +262,7 @@ solo:
 	} else if (wsi->a.vhost->socks_proxy_port) {
 		lwsl_client("Sending SOCKS Greeting\n");
 		ads = wsi->a.vhost->socks_proxy_address;
-		port = wsi->a.vhost->socks_proxy_port;
+		port = (int)wsi->a.vhost->socks_proxy_port;
 #endif
 	} else {
